\name{extract_aes_param}
\alias{extract_aes_param}
\title{Build a list of aes and param by comparing values in the droplist and default and only using those different from default...}
\usage{extract_aes_param(list_aes)}
\description{Build a list of aes and param by comparing values in the droplist and default and only using those different from default}
\value{list of list param and aes containing the values to be used in the layers}
\author{cedric}
\arguments{\item{list_aes}{:the list of defautl geom and params written in .RglobalEnv when selecting stat or geom in aes_frame}}
