\name{hsw}
\alias{hsw}
\title{handler allowing the switch between stat and geom when button geom<>stat is triggered...}
\usage{hsw(h, ...)}
\description{handler allowing the switch between stat and geom when button geom<>stat is triggered}
\author{cedric}
\arguments{\item{h}{}
\item{...}{}}
